+++ NEWS.md
@@ -1,4 +1,5 @@
 # disk.frame 0.2.2
+* experimental group-by framework!
 * bug fixes for data.table trigger by integration with tidyfast
 
 # disk.frame 0.2.1